move angle errors
